Leader Election - Passing first test
Need to clean up some of this code, and handle the other cases
but lets merge in the first part of the leader election.

All rafts start up and wait for a randomized election Timeout.
After such, they hold an election and vote for themselves.
When one raft, in candidate status, received a majority of votes - it becomes the leader.
The leader raft sends out AppendEntries heatbeats to all the other rafts solidifying its
leadership status and term.