Handle election timeout before RPC timeout
Let's say that the election timeout is about to expire,
however we are waiting on a lossy RPC from the server.
We can handle both events with a select statement to
make sure we stay on time for elections.

Still seeing some errors on Lossy test case when
running long tests.