3B: Passing all test cases
Fix multiple issues with fast backup and iterating
over the appendRPC entries in-order.
One caveat to processing the entries in the loop
prevLogIndex ... prevLogIndex + len(Entries) there
needs to be a way to determine if they are stale as
prevLogIndex has already been updated.
To determine this, we check if the cmd of the index is nil,
as the entries does not contain this index calculated by
an updated prevLogIndex. TBD:
There may be some issues with overlapping Start() using this
approach.