3B: Fix the final issue before starting 3C :D
With this commit, the 3B test cases all pass pretty consistently.
There may be some bug lingering, as sometimes the first commit or
the concurrency test fails. But, the logs are all aligning properly
now over time so it is safe to say I can move on to 3C.

As for this commit, I am fixing one main issue where the newEntries
was not getting reset within the loop. So if we send an incomplete
list of entries, and loop back over for more, we can end up sending
entries we did not mean to.