misc: change log from map to slice
On my first implementation, I decided on using a "map" which is the
Go structure for a hashmap.

For raft, which is an incremental log mechansim, a more typical array
(Go's "log") works better. We can take index slices from beginning or to end.
There is always a guarantee that each index is filled with data.